requirements.txt
fastapi
uvicorn
asyncpg
databases
python-dotenv
pydantic
pydantic-settings
email-validator
python-multipart
python-jose[cryptography]
passlib[bcrypt]

aiosqlite
